Made snagging slightly less common
Increased the chance for the bot to snag something if a user doesn't
have a quote yet.
Decreased overall chance of the bot to snag a quote.
Fixed a bug where the bot would throw an SQL error if it tried to add a
ping for a user if they changed their login, which isn't yet on the
users table.
Added a bunch of inside jokes